% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{addAuthKey}
\alias{addAuthKey}
\title{Add Authentication Key}
\usage{
addAuthKey(server_handle, key)
}
\arguments{
\item{server_handle}{External pointer from runServer(blocking=FALSE, auth=TRUE)}

\item{key}{Character string, the API key to add}
}
\value{
Invisible TRUE
}
\description{
Add an API key to the authentication system
}
