% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{as.data.frame.server_list}
\alias{as.data.frame.server_list}
\title{Convert server list to data frame}
\usage{
\method{as.data.frame}{server_list}(x, ...)
}
\arguments{
\item{x}{a server_list object}

\item{...}{additional arguments (ignored)}
}
\value{
a data.frame with server information
}
\description{
Convert server list to data frame
}
