% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{clearAuthKeys}
\alias{clearAuthKeys}
\title{Clear All Authentication Keys}
\usage{
clearAuthKeys(server_handle)
}
\arguments{
\item{server_handle}{External pointer from runServer(blocking=FALSE, auth=TRUE)}
}
\value{
Invisible TRUE
}
\description{
Remove all API keys from the authentication system
}
