% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{createFileLogHandler}
\alias{createFileLogHandler}
\title{Create file log handler}
\usage{
createFileLogHandler(fd, logfile = tempfile("goserveR_", fileext = ".log"))
}
\arguments{
\item{fd}{file descriptor for log pipe}

\item{logfile}{path to log file}
}
\value{
external pointer to log handler
}
\description{
Create file log handler
}
