% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{createSilentLogHandler}
\alias{createSilentLogHandler}
\title{Create silent log handler (no-op)}
\usage{
createSilentLogHandler(fd)
}
\arguments{
\item{fd}{file descriptor for log pipe}
}
\value{
external pointer to log handler
}
\description{
Create silent log handler (no-op)
}
