% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{.create_default_log_handler}
\alias{.create_default_log_handler}
\title{Create default console log handler}
\usage{
.create_default_log_handler(fd)
}
\arguments{
\item{fd}{file descriptor for log pipe}
}
\value{
external pointer to log handler
}
\description{
Create default console log handler
}
