% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{.default_log_callback}
\alias{.default_log_callback}
\title{Default log callback - writes to console}
\usage{
.default_log_callback(handler, message, user)
}
\arguments{
\item{handler}{external pointer to handler}

\item{message}{log message}

\item{user}{user data (unused)}
}
\description{
Default log callback - writes to console
}
