% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{isRunning}
\alias{isRunning}
\title{isRunning
Check if a background server is still running}
\usage{
isRunning(handle)
}
\arguments{
\item{handle}{external pointer returned by runServer(blocking=FALSE)}
}
\value{
logical, TRUE if server is running, FALSE otherwise
}
\description{
isRunning
Check if a background server is still running
}
\examples{
\dontrun{
h <- runServer(dir = ".", addr = "127.0.0.1:8080", blocking = FALSE)
isRunning(h) # TRUE
shutdownServer(h)
isRunning(h) # FALSE
}
}
