% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{listAuthKeys}
\alias{listAuthKeys}
\title{List Authentication Keys}
\usage{
listAuthKeys(server_handle)
}
\arguments{
\item{server_handle}{External pointer from runServer(blocking=FALSE, auth=TRUE)}
}
\value{
Character vector of current API keys
}
\description{
Get all current API keys in the authentication system
}
