% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{print.server_info}
\alias{print.server_info}
\title{Print method for individual server_info objects}
\usage{
\method{print}{server_info}(x, index = NULL, ...)
}
\arguments{
\item{x}{a server_info object}

\item{index}{server index number for display}

\item{...}{additional arguments (ignored)}
}
\description{
Print method for individual server_info objects
}
