% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{print.server_list}
\alias{print.server_list}
\title{Print method for server_list objects}
\usage{
\method{print}{server_list}(x, ...)
}
\arguments{
\item{x}{a server_list object}

\item{...}{additional arguments (ignored)}
}
\description{
Print method for server_list objects
}
