% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{registerLogHandler}
\alias{registerLogHandler}
\title{Register a log handler for a file descriptor}
\usage{
registerLogHandler(fd, callback, user = NULL)
}
\arguments{
\item{fd}{file descriptor to monitor}

\item{callback}{R function to call when data is available}

\item{user}{user data passed to callback}
}
\value{
external pointer to log handler
}
\description{
Register a log handler for a file descriptor
}
