% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{removeAuthKey}
\alias{removeAuthKey}
\title{Remove Authentication Key}
\usage{
removeAuthKey(server_handle, key)
}
\arguments{
\item{server_handle}{External pointer from runServer(blocking=FALSE, auth=TRUE)}

\item{key}{Character string, the API key to remove}
}
\value{
Invisible TRUE
}
\description{
Remove an API key from the authentication system
}
