% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{removeLogHandler}
\alias{removeLogHandler}
\title{Remove a log handler}
\usage{
removeLogHandler(handler)
}
\arguments{
\item{handler}{external pointer to log handler}
}
\value{
logical indicating success
}
\description{
Remove a log handler
}
