% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{ShutdownServer}
\alias{ShutdownServer}
\title{ShutdownServer (advanced/manual use)
Shutdown a server (C-level, advanced)}
\usage{
ShutdownServer(handle)
}
\description{
ShutdownServer (advanced/manual use)
Shutdown a server (C-level, advanced)
}
