% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{StartServer}
\alias{StartServer}
\title{StartServer (advanced/manual use)
Start a server (C-level, advanced)}
\usage{
StartServer(dir, addr, prefix, blocking)
}
\description{
StartServer (advanced/manual use)
Start a server (C-level, advanced)
}
