% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{summary.server_list}
\alias{summary.server_list}
\title{Summary method for server_list objects}
\usage{
\method{summary}{server_list}(object, ...)
}
\arguments{
\item{object}{a server_list object}

\item{...}{additional arguments (ignored)}
}
\description{
Summary method for server_list objects
}
