% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{listServers}
\alias{listServers}
\title{listServers
List all running background servers}
\usage{
listServers()
}
\value{
a list of server info
}
\description{
listServers
List all running background servers
}
