% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{listServers}
\alias{listServers}
\title{listServers
List all running background servers with detailed information}
\usage{
listServers()
}
\value{
a server_list S3 object containing server information
}
\description{
listServers
List all running background servers with detailed information
}
