% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goServeR.R
\name{shutdownServer}
\alias{shutdownServer}
\title{shutdownServer
Shutdown a background server}
\usage{
shutdownServer(handle)
}
\arguments{
\item{handle}{external pointer returned by runServer(blocking=FALSE)}
}
\description{
shutdownServer
Shutdown a background server
}
